Fix for Students Step 1

+++ GameLibrary/Controllers/BoardGameController.cs
@@ -1,6 +1,8 @@
-﻿using GameLibrary.Data;
-using DataModel;
+﻿using DataModel;
+using GameLibrary.Data;
+using DataModel; 
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 
 namespace GameLibrary.Controllers
@@ -14,18 +16,23 @@ namespace GameLibrary.Controllers
         }
         public ActionResult Index()
         {
+            //ToDo add publisher to the BoardGameModel
             return View(_context.BoardGames
-                .Include(p => p.Publishers).ToList());
+                .OrderBy(b => b.PublishersId) // order by Id so that all books by the same publisher will be together
+                .Include(p => p.Publishers)
+                .ToList());
         }
         public ActionResult Create()
         {
-            return View();
+            BoardGameModel boardGame = new BoardGameModel();
+            return View(boardGame);
         }
         [HttpPost]
         public ActionResult Create(BoardGameModel boardGame)
         {
             if (ModelState.IsValid)
             {
+                //ToDo uses NewPublisher for the user to enter a publisher new
                 boardGame.PublishersId = GetPublisher(boardGame.NewPublisher);
                 _context.BoardGames.Add(boardGame);
                 _context.SaveChanges();
@@ -35,15 +42,15 @@ namespace GameLibrary.Controllers
         }
         public ActionResult Delete(int id)
         {
-            if (id == null)
+            if (id == 0)
             {
-                return NotFound();
+                return Problem($"BoardGame Id is null");
             }
             var game = _context.BoardGames.FirstOrDefault(
                 g => g.Id == id);
             if (game == null)
             {
-                return NotFound();
+                return Problem($"BoardGame Id {id} was not found");
             }
             return View(game);
         }
@@ -57,16 +64,33 @@ namespace GameLibrary.Controllers
                 return Problem("Entity set 'GameContext.BoardGames is null");
             }
             var game = _context.BoardGames.Find(id);
-            _context.BoardGames.Remove(game);
-            _context.SaveChanges();
+            if (game is not null) //ToDo: Added A Null Check to prevent errors
+            {
+                _context.BoardGames.Remove(game);
+                _context.SaveChanges();
+            }
+            else
+            {
+                return Problem($"Unable to delete BoardGames id {id}");
+            }
             return RedirectToAction("Index");
         }
         public ActionResult Edit(int id)
-        {      
+        {
+            BoardGameModel? game = _context.BoardGames
+                                    .Include(p=>p.Publishers)
+                                    .Where(b => b.Id == id)
+                                    .FirstOrDefault();
+            if (game is not null)
+            {
+                game.NewPublisher = game.Publisher;
+                return View(game);
+            }
+
             return View();
         }
         [HttpPost]
-        public ActionResult Edit(int id, BoardGameModel game)
+        public ActionResult Edit(BoardGameModel game)
         {
             if (ModelState.IsValid)
             {
@@ -81,31 +105,38 @@ namespace GameLibrary.Controllers
         {
             if (id == 0)
             {
-                return NotFound();
+                return Problem("Entity set 'GameContext.BoardGames is null");
             }
             var game = _context.BoardGames
                 .Where(b => b.Id == id)
                 .Include(b => b.Publishers)
                 .FirstOrDefault();
-
+            if (game is null)
+            {
+                return Problem($"Unable to find BoardGames id {id}");
+            }
             return View(game);
         }
-        //This method gets the publishers Id if it the publisher exists already,
-        //if it does not exist it creates a new publisher
+
+        //ToDo Gets the Publisher ID even if it has to add the publisher
         private int GetPublisher(string publisher)
         {
             PublisherModel? pub = null;
+
             pub = _context.Publishers
-                .Where(p => p.Name.ToLower() == publisher.ToLower())
-                .FirstOrDefault();
-            if (pub == null)
+                   .Where(p => p.Name.ToLower() == publisher)
+                   .FirstOrDefault();
+            if (pub is null)
             {
-                pub = new PublisherModel { Name = publisher };
+                pub = new PublisherModel
+                {
+                    Name = publisher
+                };
                 _context.Add(pub);
-                _context.SaveChanges(true);
+                _context.SaveChanges();
             }
             return pub.Id;
-
         }
+
     }
 }
